+++ game-falloutlondon/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Future Changes (NOT IMPLEMENTED YET):
 
-- Find out how to change modtype for falloutlondon mod automatically (? set from modtype ?)
+- Find out how to change modtype for falloutlondon mod automatically (? set from registerModType ?)
 - Test INI code
 - Add check for if INI files already updated