+++ game-falloutlondon/CHANGELOG.md
@@ -3,8 +3,7 @@
 Future Changes (NOT IMPLEMENTED YET):
 
 - Find out how to change modtype for falloutlondon mod automatically (? set from registerModType ?)
-- Test INI code
-- Add check for if INI files already updated
+- Add working check for if INI files already updated (existing doesnt seem to work)
 
 ## [0.1.0] - 2025-06-24