+++ game-falloutlondon/CHANGELOG.md
@@ -3,8 +3,12 @@
 Future Changes (NOT IMPLEMENTED YET):
 
 - Find out how to change modtype for falloutlondon mod automatically (? set from registerModType ?)
-- Add working check for if INI files already updated (existing doesnt seem to work)
+- Add working check for if INI files already updated (existing check doesnt seem to work)
 
-## [0.1.0] - 2025-06-24
+## [0.1.1] - 2025-06-24
+
+- Minor improvements
+
+## [0.1.0] - 2025-06-23
 
 - Initial release